requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
tensorflow==2.17.0
pillow==11.1.0
numpy==1.26.4
h5py==3.12.1